Add an unsafe interface to register statics.
The Library macro create an empty tuple pointer storage to reference the space
which contains the value of all the statics value sused during the compilation.

This implies that the library cannot use the untyped tuple type to derive the
size, using mem::size_of, to assert its size when registering the memory of
statics.